Moved Image Uploading into its own file
- All image uploading code moved to its own file
- Expanded error logic
- Client updated for fallback image
